+++ src/Model/Launcher/Launcher.php
@@ -11,7 +11,8 @@
 
 namespace ExtensionsForGrifus\Modules\CustomImagesGrifus\Model\Launcher;
 
-use Eliasis\Model\Model;
+use Eliasis\Model\Model,
+	Eliasis\Module\Module;
     
 /**
  * Module main model.
@@ -28,5 +29,55 @@ class Launcher extends Model {
     public function deletePostMeta() {
 
         delete_post_meta_by_key('custom_images_grifus'); 
+    }
+
+    /**
+     * Set database module options.
+     * 
+     * @since 1.0.1
+     *
+     * @return void
+     */
+    public function addOptions() {
+        
+        $slug = Module::CustomImagesGrifus()->get('slug');
+
+        if (!get_option($slug . '-replace-when-add')) {
+
+            add_option($slug . '-replace-when-add', 1);
+        }
+    }
+
+    /**
+     * Get database module options.
+     *
+     * @since 1.0.1
+     *
+     * @uses get_option()
+     *
+     * @return array
+     */
+    public function getOptions() {
+
+        $slug = Module::CustomImagesGrifus()->get('slug');
+
+        return [
+
+            'replace-when-add' => get_option($slug . '-replace-when-add')
+        ];
+    }
+
+    /**
+     * Delete database module options.
+     * 
+     * @since 1.0.1
+     *
+     * @return void
+     */
+    public function deleteOptions() {
+
+        $slug = Module::CustomImagesGrifus()->get('slug');
+
+        delete_option($slug . '-replace-when-add', true);
     }
 }